refined vbpr implementation
