+++ clayrs/recsys/__init__.py
@@ -5,8 +5,8 @@ from . import graphs
 from .content_based_algorithm import *
 from .graph_based_algorithm import *
 from .graphs import *
+from .visual_based_algorithm import *
 from .recsys import ContentBasedRS, GraphBasedRS
 from .partitioning import KFoldPartitioning, HoldOutPartitioning, BootstrapPartitioning
 from .methodology import TestRatingsMethodology, TestItemsMethodology, TrainingItemsMethodology, AllItemsMethodology
 from .experiment import ContentBasedExperiment, GraphBasedExperiment
-from .visual import VBPR
