+++ clayrs/recsys/visual_based_algorithm/__init__.py
@@ -1,0 +1,1 @@
+from .vbpr import *
