+++ clayrs/recsys/visual_based_algorithm/vbpr/__init__.py
@@ -1,0 +1,1 @@
+from .vbpr_algorithm import VBPR
